+++ BUILD.bazel
@@ -703,16 +703,6 @@ cc_test(
     ],
 )
 
-cc_test(
-    name = "gcs_server_test",
-    srcs = glob(["src/ray/gcs/gcs_server/test/gcs_server_test.cc"]),
-    copts = COPTS,
-    deps = [
-        ":gcs_server_lib",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
 cc_library(
     name = "object_manager",
     srcs = glob([
