+++ src/ray/rpc/gcs_server/gcs_rpc_client.h
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef RAY_RPC_GCS_RPC_CLIENT_H
+#define RAY_RPC_GCS_RPC_CLIENT_H
 
 #include <thread>
 
@@ -95,3 +96,5 @@ class GcsRpcClient {
 
 }  // namespace rpc
 }  // namespace ray
+
+#endif // RAY_RPC_GCS_RPC_CLIENT_H
